finish gameinit and merge Initial.java into ServerGameController.java
