+++ src/ncu/sw/gameServer/Main.java
@@ -4,8 +4,8 @@ public class Main {
     public static void main(String[] args) {
         System.out.println("server main");
         //launch(args);
-      // ServerGameController server = new ServerGameController(100,20,20);
-        Initial a = new Initial(100,30,80);
+        ServerGameController gameController = new ServerGameController(100,20,20);
+
         MultiServer server = new MultiServer(9000);
         new Thread(server).start();
     }
